pyproject.toml
[project]
name = "deep-research-agent"
version = "0.1.0"
description = "Deep Research Agentic AI - Comprehensive business idea research and analysis system"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.2.4",
    "tavily-python>=0.7.10",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "asyncio",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

